Fatima_Sitah

+++ restaurants/views.py
@@ -19,7 +19,8 @@ def signup(request):
         "form":form,
     }
     return render(request, 'signup.html', context)
-
+def no_access(request):
+    return render(request, 'no_access.html')
 def signin(request):
     form = SigninForm()
     if request.method == 'POST':
@@ -59,6 +60,8 @@ def restaurant_detail(request, restaurant_id):
     return render(request, 'detail.html', context)
 
 def restaurant_create(request):
+    if request.user.is_anonymous:
+        return redirect('signin')
     form = RestaurantForm()
     if request.method == "POST":
         form = RestaurantForm(request.POST, request.FILES)
@@ -75,6 +78,8 @@ def restaurant_create(request):
 def item_create(request, restaurant_id):
     form = ItemForm()
     restaurant = Restaurant.objects.get(id=restaurant_id)
+    if not(request.user.is_staff or request.user == restaurant.owner):
+        return redirect('no-access')
     if request.method == "POST":
         form = ItemForm(request.POST)
         if form.is_valid():
@@ -90,6 +95,8 @@ def item_create(request, restaurant_id):
 
 def restaurant_update(request, restaurant_id):
     restaurant_obj = Restaurant.objects.get(id=restaurant_id)
+    if not(request.user.is_staff or request.user == restaurant_obj.owner):
+        return redirect('no-access')
     form = RestaurantForm(instance=restaurant_obj)
     if request.method == "POST":
         form = RestaurantForm(request.POST, request.FILES, instance=restaurant_obj)
@@ -104,5 +111,7 @@ def restaurant_update(request, restaurant_id):
 
 def restaurant_delete(request, restaurant_id):
     restaurant_obj = Restaurant.objects.get(id=restaurant_id)
+    if not (request.user.is_staff):
+        return redirect('no-access')
     restaurant_obj.delete()
     return redirect('restaurant-list')
